More work on available expressions
Found that Data.Set does not implement == as
expected -- which caused available expression
analysis to loop, because facts did not reach
a fixed point.

Added "LitM" expression to deal with numerics. Instead
of treating them as "variables", they become a monadic
bind:

  f 1
 ==>
  v <- num 1
  f @ v

Still not ideal but a little better.

Code cleanup throughout.