Fixing up live variables, other bug fixes.
Liveness with primitives not working quite right - now
fixed. Top-level user-defined and primitive names kept
separate. Primitives and top-level names now correctly
eliminated.

Successor definition for MIL not quite right - a block
that returns a closure is no the predecessor of the
destination stored in the closure! Fixed.

Code cleanup as usual ...