Change strategy for Case expressions
Case expressions previously compiled using
"join points". Now, compile case expression
and arms in a separate block. Use Goto to get the
result of the case expresssions. Changed to fix
some problems with free variables and case expressions,
specifically:

  (case y of ...) x

Previous strategy was not able to recognize that x was free and
needed to be included in join point function.